fixing delimiters and consolidating into few
